ci: ensure test coverage without a default target
This PR adds 'tools/missing' and a 'missing' Makefile target useful
for finding packages that are not tested by the test-pkgs matrix in
CI via tests-core.yaml.

In comparison, the Circle config tests all packages by default in the
'tests-other' job unless explicitly tested in another job. This leads
to an accumulation of packages and a long running, often failing job.

In GHA we split out packages a lot more than in Circle, because disk
/ netowrk performance is extremely poor. Mapping out the work to many
runners helps get the runtime down.